+++ src/routes/(landing-pages)/career/builderio/+page.server.ts
@@ -1,7 +1,0 @@
-import { redirect } from '@sveltejs/kit';
-
-export function load() {
-	throw redirect(301, '/career/builder');
-}
-
-export const prerender = true;